oof_unet/pyproject.toml
[project]
name = "oof_unet"
version = "1.0"
requires-python = ">=3.11"
description = "FCCSS is a Python Package to train UNet architectures to predict out of field dosemaps"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "EL AICHI Mohammed", email = "mohammed.EL-AICHI@gustaveroussy.fr" },
    { name = "Institut Gustave Roussy, Unit U 1030, Radiotherapy department" }
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Computer vision",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
keywords = [
    'deep learning',
    'image segmentation',
    'Out Of Field Dose estimation'
]
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "pandas",
    "antspyx"
    ]

[project.urls]
homepage = "https://github.com/maichi98/OOF_UNET"
repository = "https://github.com/maichi98/OOF_UNET.git"


[build-system]
requires = ["setuptools>=69.5.1"]
build-backend = "setuptools.build_meta"

[tool.codespell]
skip = '.git,*pdf,*.svg'